Insert empty line between suite and alternative branch after def/class
When there is a function or class definition at the end of a suite followed by the beginning of an alternative block, we have to insert a single empty line between them.

In the if-else-statement example below, we insert an empty line after the `foo` in the if-block, but none after the else-block `foo`, since in the latter case the enclosing suite already adds empty lines.

```python
if sys.version_info >= (3, 10):
    def foo():
        return "new"
else:
    def foo():
        return "old"
class Bar:
    pass
```

To do so, we track whether the current suite is the last one in the current statement with a new option on the suite kind.

Fixes #12199