Display fixes by default (#7352)
Use the existing `printer::Flags::SHOW_FIX_DIFF` to enable printing
fixable diffs in `check` mode automatically. Enable it whenever the
output format is `Full`, but keep a distinct flag for it (a) because
output format and whether to print diffs are really separate concerns,
(b) because it makes it easy to make changes to what sets the flag to
true if/as that is needful, and (c) because it minimally perturbs the
rest of the code.

This does not introduce *new* tests, but does update the existing tests
to account for the new default output.