Only show safe fixes in `check` output for now
There are some open design questions about how the calls to action
should work for unsafe and display-only fixes, so leave them aside for
now. This hard-codes that behavior into the `Display` implementation,
but with a tracking issue for moving that up to the `TextEmitter`, and
also includes a TODO with a tracking issue for the design of the CTAs.